requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
django-cors-headers==3.1.1
httpie==1.0.2
idna==2.8
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.3
#psycopg2==2.8.4
jsonfield==2.0.2
xlwt==1.3.0

